+++ tests/modelProgram/app/_ui/scripts/helloworld.py
@@ -1,0 +1,1 @@
+print("hello world")
