+++ tests/modelProgram/data/1_gus.csv
@@ -1,2 +1,3 @@
 lat,lon,l_motor_speed,r_motor_speed,net_velocity,net_accel,voltage,1_temp,2_temp,3_temp,1_phidg_v,1_phidg_c,2_phidg_v,2_phidg_c,speed,average_temp,heading,battery,status,mode,name,ip_address,network_status,temp_sensor_1,temp_sensor_2,temp_sensor_3,volt_5v_line_sensor,volt_killsig_line_sensor,volt_24v_line_sensor,volt_curr_24v_line_sensor,volt_curr_thr_line_sensor,battery_status,lidar_sensor,video_feed,teensy_status,thruster_status
-32.70465547153673,-117.22840547810463,45.954620217232325,22.85987577733029,34.927640266170975,82.50062234034063,90.86639750854376,81.14014313997461,90.99249295126164,65.92117434813738,40.25419132969448,5.4001594144627685,95.76809725024032,76.27937960325387,54.74551317503767,54.47941639594368,283.0104051834354,31.25882756664473,97.64626753630223,0.05992720213424674,74.41362674110292,46.26108190155156,77.30821325320609,30.698564552884434,52.66099609103338,68.2721941320659,34.437984397621754,85.65746184422922,63.82280959003636,65.68646008130833,50.583178708781595,99.3874133028758,82.77255486211098,39.61648502166501,87.66811875692531,68.66275252225735
+32.70384645454236,-117.22899326335693,45.954620217232325,22.85987577733029,34.927640266170975,82.50062234034063,90.86639750854376,81.14014313997461,90.99249295126164,65.92117434813738,40.25419132969448,5.4001594144627685,95.76809725024032,76.27937960325387,47.171165158103264,15.43360696034889,53.25357724641534,5.280258376211588,10.985181725111904,63.07397527614974,24.695545468947522,55.24288501686151,69.8350401320554,42.5549264933239,69.02368191277787,51.61492564943636,3.3370839995685353,27.725829385452037,18.89324449751558,22.683151225822662,44.77539212402548,83.75382836979833,21.568935496549635,66.90194399024345,38.38674356449718,57.7726383555245
+32.70384645454236,-117.22899326335693,,,,,,,,,,,,,47.171165158103264,15.43360696034889,53.25357724641534,5.280258376211588,10.985181725111904,63.07397527614974,24.695545468947522,55.24288501686151,69.8350401320554,42.5549264933239,69.02368191277787,51.61492564943636,3.3370839995685353,27.725829385452037,18.89324449751558,22.683151225822662,44.77539212402548,83.75382836979833,21.568935496549635,66.90194399024345,38.38674356449718,57.7726383555245
