+++ tests/modelProgram/data/2_gus.csv
@@ -1,2 +1,3 @@
 lat,lon,l_motor_speed,r_motor_speed,net_velocity,net_accel,voltage,1_temp,2_temp,3_temp,1_phidg_v,1_phidg_c,2_phidg_v,2_phidg_c,speed,average_temp,heading,battery,status,mode,name,ip_address,network_status,temp_sensor_1,temp_sensor_2,temp_sensor_3,volt_5v_line_sensor,volt_killsig_line_sensor,volt_24v_line_sensor,volt_curr_24v_line_sensor,volt_curr_thr_line_sensor,battery_status,lidar_sensor,video_feed,teensy_status,thruster_status
-32.704552088309185,-117.22842185239926,92.44355212654888,17.6126936970058,90.16359589740462,79.40065610366977,52.730515813919055,27.874721290177863,60.52723504096952,2.625735029840448,22.971685894044903,32.940624398025776,85.71233958939929,79.93442463863435,62.393419076330936,86.03163912702195,320.9126499751024,21.306523824018242,58.24738107252162,68.89334464792354,73.03009065486536,62.82356360444595,2.5576631024317176,78.48672689805008,47.878206930937594,27.83934105479885,58.9934283672398,44.526070154609506,47.923408470287995,74.77865785442944,55.96683548969331,35.693028043307194,67.06762712816375,70.34769334479033,63.869143694384114,8.88489419569467
+32.70380894348371,-117.22909098300562,92.44355212654888,17.6126936970058,90.16359589740462,79.40065610366977,52.730515813919055,27.874721290177863,60.52723504096952,2.625735029840448,22.971685894044903,32.940624398025776,85.71233958939929,79.93442463863435,22.46075148187202,39.67948070228844,130.0087430512027,82.88886325657376,13.037834234078572,27.11794610857169,2.3474719259655163,60.89333576122932,97.0700516495703,89.64606777241362,66.64521929585217,13.6536137101129,79.68244447759395,28.48732575023346,76.23528263220926,24.441942494403847,9.11716036947351,60.8018825113407,6.95185633657659,80.41257751425506,92.07417804893463,26.434601441710292
+32.70380894348371,-117.22909098300562,,,,,,,,,,,,,22.46075148187202,39.67948070228844,130.0087430512027,82.88886325657376,13.037834234078572,27.11794610857169,2.3474719259655163,60.89333576122932,97.0700516495703,89.64606777241362,66.64521929585217,13.6536137101129,79.68244447759395,28.48732575023346,76.23528263220926,24.441942494403847,9.11716036947351,60.8018825113407,6.95185633657659,80.41257751425506,92.07417804893463,26.434601441710292
