+++ tests/modelProgram/data/3_gus.csv
@@ -1,2 +1,4 @@
 lat,lon,_motor_speed,r_motor_speed,net_velocity,net_accel,voltage,1_temp,2_temp,3_temp,1_phidg_v,1_phidg_c,2_phidg_v,2_phidg_c,l_motor_speed,speed,average_temp,heading,battery,status,mode,name,ip_address,network_status,temp_sensor_1,temp_sensor_2,temp_sensor_3,volt_5v_line_sensor,volt_killsig_line_sensor,volt_24v_line_sensor,volt_curr_24v_line_sensor,volt_curr_thr_line_sensor,battery_status,lidar_sensor,video_feed,teensy_status,thruster_status
-32.70445098300563,-117.2284489434837,32.56670406025014,52.730410688186105,95.09959115773864,48.2165634083206,65.01800947505171,75.00548311845465,85.6815795083452,44.39327854185995,24.576986272006497,89.90880724548501,86.5755101487587,86.89766559935865,70.4214123646583,24.161812857633038,13.225461413448802,82.60851699348652,10.085333395246154,44.74167261369008,9.140591811718412,77.58069200643179,76.68594461440514,4.163889492971862,32.20029092066189,81.52264286064106,8.447007306997723,59.12520758041092,24.916152238850863,72.85709040451421,58.31223873196073,82.45764977984823,17.434091750310586,80.84201246566263,90.50835025372518,54.25793936348342,86.25771939242452
+32.70378185239927,-117.22919208830918,32.56670406025014,52.730410688186105,95.09959115773864,48.2165634083206,65.01800947505171,75.00548311845465,85.6815795083452,44.39327854185995,24.576986272006497,89.90880724548501,86.5755101487587,86.89766559935865,70.4214123646583,16.08242603096549,53.17620748230358,226.33959951318812,1.49425570505749,36.38375809871787,45.86378569471098,72.20113186295501,49.46924898211584,17.833396513112664,93.36195525129052,39.71948322873309,19.335763965741283,49.70143298626816,22.26001340345495,33.92229678431302,36.17702962078438,21.4232156386844,56.5486040669667,72.19120450670444,37.976420358811716,47.88362350091234,86.30920634046501
+32.70378185239927,-117.22919208830918,,,,,,,,,,,,,,16.08242603096549,53.17620748230358,226.33959951318812,1.49425570505749,36.38375809871787,45.86378569471098,72.20113186295501,49.46924898211584,17.833396513112664,93.36195525129052,39.71948322873309,19.335763965741283,49.70143298626816,22.26001340345495,33.92229678431302,36.17702962078438,21.4232156386844,56.5486040669667,72.19120450670444,37.976420358811716,47.88362350091234,86.30920634046501
+32.70378185239927,-117.22919208830918,,,,,,,,,,,,,,16.08242603096549,53.17620748230358,226.33959951318812,1.49425570505749,36.38375809871787,45.86378569471098,72.20113186295501,49.46924898211584,17.833396513112664,93.36195525129052,39.71948322873309,19.335763965741283,49.70143298626816,22.26001340345495,33.92229678431302,36.17702962078438,21.4232156386844,56.5486040669667,72.19120450670444,37.976420358811716,47.88362350091234,86.30920634046501
