+++ tests/modelProgram/data/4_gus.csv
@@ -1,2 +1,5 @@
 lat,lon,_motor_speed,r_motor_speed,net_velocity,net_accel,voltage,1_temp,2_temp,3_temp,1_phidg_v,1_phidg_c,2_phidg_v,2_phidg_c,l_motor_speed,speed,average_temp,heading,battery,status,mode,name,ip_address,network_status,temp_sensor_1,temp_sensor_2,temp_sensor_3,volt_5v_line_sensor,volt_killsig_line_sensor,volt_24v_line_sensor,volt_curr_24v_line_sensor,volt_curr_thr_line_sensor,battery_status,lidar_sensor,video_feed,teensy_status,thruster_status
-32.70435326335692,-117.22848645454235,87.16048358259172,10.961990511145569,1.3036921974332372,32.16932210283526,49.09655698232229,15.23794952291423,14.51176690150292,5.2088339624715925,82.4247555695942,26.24017769955813,17.002122634739404,22.31755389748654,35.20340004608335,5.800165465817497,96.10965156949989,132.99322183612935,4.692807248080899,1.8941852695139705,83.51467743588745,28.499518072587684,96.0588555934185,94.56927673050069,16.41776796685629,99.51658311957195,36.367695385731565,52.33613330994945,8.407893931064269,97.92256639806054,16.1339284372248,69.11363816824073,50.74762703807421,25.1755703972386,74.43150478619317,64.87780495680175,51.51713843754241
+32.70376547810463,-117.22929547153673,87.16048358259172,10.961990511145569,1.3036921974332372,32.16932210283526,49.09655698232229,15.23794952291423,14.51176690150292,5.2088339624715925,82.4247555695942,26.24017769955813,17.002122634739404,22.31755389748654,35.20340004608335,84.38893287446449,59.76468715811911,51.788682462831794,71.61604374174009,3.955713069647715,39.644365123283684,85.63497885459113,48.61523509417085,83.55246982891062,85.3480985505915,90.06220178719016,27.091385395402067,58.58320765305336,92.19582557019635,18.98476579154691,62.7897692472284,73.38543167714774,93.88706513129719,90.003442174948,2.4943579888361045,51.51235688073206,79.87262437409441
+32.70376547810463,-117.22929547153673,,,,,,,,,,,,,,84.38893287446449,59.76468715811911,51.788682462831794,71.61604374174009,3.955713069647715,39.644365123283684,85.63497885459113,48.61523509417085,83.55246982891062,85.3480985505915,90.06220178719016,27.091385395402067,58.58320765305336,92.19582557019635,18.98476579154691,62.7897692472284,73.38543167714774,93.88706513129719,90.003442174948,2.4943579888361045,51.51235688073206,79.87262437409441
+32.70376547810463,-117.22929547153673,,,,,,,,,,,,,,84.38893287446449,59.76468715811911,51.788682462831794,71.61604374174009,3.955713069647715,39.644365123283684,85.63497885459113,48.61523509417085,83.55246982891062,85.3480985505915,90.06220178719016,27.091385395402067,58.58320765305336,92.19582557019635,18.98476579154691,62.7897692472284,73.38543167714774,93.88706513129719,90.003442174948,2.4943579888361045,51.51235688073206,79.87262437409441
+32.70376547810463,-117.22929547153673,,,,,,,,,,,,,,84.38893287446449,59.76468715811911,51.788682462831794,71.61604374174009,3.955713069647715,39.644365123283684,85.63497885459113,48.61523509417085,83.55246982891062,85.3480985505915,90.06220178719016,27.091385395402067,58.58320765305336,92.19582557019635,18.98476579154691,62.7897692472284,73.38543167714774,93.88706513129719,90.003442174948,2.4943579888361045,51.51235688073206,79.87262437409441
