+++ tests/modelProgram/data/5_gus.csv
@@ -1,2 +1,3 @@
 lat,lon,_motor_speed,r_motor_speed,net_velocity,net_accel,voltage,1_temp,2_temp,3_temp,1_phidg_v,1_phidg_c,2_phidg_v,2_phidg_c,l_motor_speed,speed,average_temp,heading,battery,status,mode,name,ip_address,network_status,temp_sensor_1,temp_sensor_2,temp_sensor_3,volt_5v_line_sensor,volt_killsig_line_sensor,volt_24v_line_sensor,volt_curr_24v_line_sensor,volt_curr_thr_line_sensor,battery_status,lidar_sensor,video_feed,teensy_status,thruster_status
-32.70426,-117.22853397459622,65.56965694634351,17.685839339969633,77.35720498870538,27.31619348109293,50.95993932280651,43.60678689495209,86.97661783814465,64.19032706282076,79.01232050492519,62.72697039208035,99.55685306064264,28.689910446648604,8.567848101611986,95.57123330414376,26.723675422532022,112.41171593705361,9.366603968025544,50.451921652121456,11.444456475248233,25.053910786141643,40.12740300575093,18.725049474235966,47.539594442963086,93.21166932780778,11.530112206194232,83.21797125447868,2.8330747444424875,85.39918726730393,75.74475206389285,83.06981708086857,18.065891568705283,26.068839417647062,35.211187285418056,76.53435825869573,42.668034900776085
+32.70376,-117.2294,65.56965694634351,17.685839339969633,77.35720498870538,27.31619348109293,50.95993932280651,43.60678689495209,86.97661783814465,64.19032706282076,79.01232050492519,62.72697039208035,99.55685306064264,28.689910446648604,8.567848101611986,26.81391000690998,90.0699149270757,64.22166902483323,23.534013829433942,6.6723950577816655,49.95533292392671,80.79521084397969,16.596110401031062,72.85926635082033,69.56286293570687,74.40602711684421,4.506427452899919,19.812890690927844,50.87398836684556,50.444430685075794,43.89001959149468,82.48455682923371,9.802176535294915,38.59139240974094,69.02211748547938,28.21963717068212,13.541847990536048
+32.70376,-117.2294,,,,,,,,,,,,,,26.81391000690998,90.0699149270757,64.22166902483323,23.534013829433942,6.6723950577816655,49.95533292392671,80.79521084397969,16.596110401031062,72.85926635082033,69.56286293570687,74.40602711684421,4.506427452899919,19.812890690927844,50.87398836684556,50.444430685075794,43.89001959149468,82.48455682923371,9.802176535294915,38.59139240974094,69.02211748547938,28.21963717068212,13.541847990536048
